% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuIncCbnn.R
\name{cuIncCbnn}
\alias{cuIncCbnn}
\title{cuIncCbnn}
\usage{
cuIncCbnn(fit, times = times, x_test = x_test)
}
\arguments{
\item{fit}{(list) output from fitHazard function.}

\item{times}{(numeric) vector of time points at which you wish you calculate the cumulative incidence.}

\item{x_test}{(matrix) matrix containing the data you wish to predict on.}
}
\value{
(matrix) a matrix representing the cumulative incidence of each individual.
}
\description{
This function calculates the cumulative incidence using a cbnn model.
}
\examples{
library(cbnn)
library(casebase)
library(magrittr)
data<-casebase::ERSPC
data$ScrArm<-as.numeric(data$ScrArm)-1
eventVar<-"DeadOfPrCa"
timeVar<-"Follow.Up.Time"
features<-"ScrArm"
nnInput<-keras::layer_input(shape=length(features))
nnOutput<-nnInput \%>\% keras::layer_dense(units=1, use_bias = TRUE)
cbnnPrep<-prepCbnn(features, nnInput, nnOutput, data, offset=NA,timeVar,
eventVar, ratio=10, compRisk=FALSE)
fit<-fitHazard(cbnnPrep,epochs=1,batchSize=500)
times<-seq(min(data$Follow.Up.Time),max(data$Follow.Up.Time),length.out=10)
x_test<-as.matrix(data[sample(nrow(data),100),])
cumulativeIncidenceCurves<-cuIncCbnn(fit, times=times,x_test=x_test)
}
